+++ fractal_server/images/__init__.py
@@ -1,4 +1,3 @@
 from .models import SingleImage  # noqa: F401
 from .models import SingleImageTaskOutput  # noqa: F401
 from .models import SingleImageUpdate  # noqa: F401
-from .models import SingleImageWithWarnings  # noqa: F401
